Fixes issue with MySQL client writing binary encoded length as a 4 bytes int but should be a 2 bytes short - fixes #27

+++ db-async-common/src/main/scala/com/github/mauricio/async/db/util/ChannelWrapper.scala
@@ -82,7 +82,7 @@ class ChannelWrapper( val buffer : ChannelBuffer ) extends AnyVal {
       buffer.writeByte( length.asInstanceOf[Byte])
     } else if (length < 65536L) {
       buffer.writeByte(252)
-      buffer.writeInt(length.asInstanceOf[Int])
+      buffer.writeShort(length.asInstanceOf[Int])
     } else if (length < 16777216L) {
       buffer.writeByte(253)
       writeLongInt(length.asInstanceOf[Int])
